Move access tests to unit folder
